Create new Knife script and knife commontrait, refactor new InventoryApply to Knife script.
