+++ UnityProject/Assets/Resources/ScriptableObjects/Traits/CommonTraitsSingleton.asset
@@ -35,3 +35,4 @@ MonoBehaviour:
   Cable: {fileID: 11400000, guid: e39c048ec864b47a495a40f9c76af764, type: 2}
   Welder: {fileID: 11400000, guid: aaed69546e98a47d8b488a046d00ebe7, type: 2}
   Shovel: {fileID: 11400000, guid: 35ecd7ae92ffc9f4d9e62613c2ad8cf8, type: 2}
+  Knife: {fileID: 11400000, guid: 5beae98557a654ea6acad86c309950af, type: 2}
