+++ UnityProject/Assets/Scripts/Items/Others/Knife.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58efcc097778a3646b453ece41fb2ab1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
