+++ UnityProject/Assets/Scripts/Items/Traits/CommonTraits.cs
@@ -32,4 +32,5 @@ public class CommonTraits : SingletonScriptableObject<CommonTraits>
 	public ItemTrait Cable;
 	public ItemTrait Welder;
 	public ItemTrait Shovel;
+	public ItemTrait Knife;
 }
